# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/friends/request/my-friend-request": {
            "get": {
                "tags": [
                    "Personal Friend Request"
                ],
                "summary": "Personal Friend Request",
                "description": "Personal Friend Request",
                "operationId": "personal friend request",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthennticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/friends/request/add": {
            "post": {
                "tags": [
                    "Add new Friend Request"
                ],
                "summary": "Add new Friend Request",
                "description": "Add new Friend Request",
                "operationId": "add new friend request",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "request_to"
                                ],
                                "properties": {
                                    "text": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful created comment",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthennticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/friends/request/confirm/{friendrequest}": {
            "put": {
                "tags": [
                    "Confirm new Friend Request"
                ],
                "summary": "Confirm new Friend Request",
                "description": "Confirm new Friend Request",
                "operationId": "confirm new friend request",
                "responses": {
                    "201": {
                        "description": "successful confirm friend request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthennticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/friends/request/decline/{friendrequest}": {
            "put": {
                "tags": [
                    "Decline Friend Request"
                ],
                "summary": "Decline Friend Request",
                "description": "Decline Friend Request",
                "operationId": "decline friend request",
                "responses": {
                    "201": {
                        "description": "successful decline friend request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthennticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/friends/request/cancel/{friendrequest}": {
            "delete": {
                "tags": [
                    "Cancel Friend Request"
                ],
                "summary": "Cancel Friend Request",
                "description": "Cancel Friend Request",
                "operationId": "cancel friend request",
                "responses": {
                    "201": {
                        "description": "successful cancel friend request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthennticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}